+++ source/weapon.cpp
@@ -49,9 +49,9 @@ Weapon::~Weapon() = default;
 
 RollSet &Weapon::get_damage() { return m_damage; }
 
-int Weapon::get_minCriticalDice() const { return m_minCriticalDice; }
+unsigned short Weapon::get_minCriticalDice() const { return m_minCriticalDice; }
 
-int Weapon::get_criticalMultiplier() const
+unsigned short Weapon::get_criticalMultiplier() const
 {
     return m_criticalMultiplier;
 }
